Detail the plugin-system design
> Why Changes/Additions/Deletions?

plugin-system.md & images (.draw/.svg): Something like the plugin-system that is
primarily going to be taking function calls as inputs isn't going to be easy to
understand just from reading the code and neither is it something that can just
be explained over a short comment on an Issue/PR. A clear explanation of how
the plugin-system works will help devs (other than myself) when contributing to
the project, especially in its initial days.

README.md: The readme is a quick intro to the project, at this early stage in
development, it is sensible to point potensial contributors directly to what
matters most.

> How it helps?

plugin-system.md & images (.draw/.svg): A clear overview of the plugin-system. If
your not dumb, you should understand how it works.

README.md: Now tells you to update yourself on the plugin-system, git
practices (that this project expects out of you) and all of the supplimentary docs
you'll have to write up and its formatting details right off the bat. At least you
can't say we didn't tell you.