Auto-install suggested mypy-stubs
> Why Changes/Additions/Deletions?

code-standards-check.yml: If mypy can’t find stubs for a third-party
library, and it knows that stubs exist for the library, it throws an
error causing the check to fail even if it's technically correct.

> How it helps?

code-standard-check.yml: mypy will now automatically install suggested
stubs before running the actual check. It's a little slower but it works.

> Notes

https://mypy.readthedocs.io/en/stable/running_mypy.html#library-stubs-not-installed