Document style-guide for diagrams
> Why Changes/Additions/Deletions?

all changes: diagrams are a crucial part of our documentation system
but, as of now contributors are just told to "create the necessary
diagrams". That doesn't usually mean uniformity of diagrams. The styling
of diagrams should be specified if it is to be made uniform without
having to specify the styling each and every time (like is being done
now)

> How it helps?

all changes: explicitly specifies styling of diagrams