Implemented the Song object
> Why Changes/Additions/Deletions?

song_obj.py: Having a dict/tuple or some other generic data-structure to
contain the details of a song is troublesome either because most python
data-structure's are mutable or because you'd need to remember dict keys
or tuple value orders. It'd also mean that there isn't necessarily going
to be input validation. Since, individual songs are at the core of
spotDL, it would make sense to pay attention to how data is "bundled" as
this data will be accessed all over the place.

> How it helps?

song_obj.py: Storing data in a class provides a few advantages that
address the very reason behind creation of said class:

  1. You can validate inputs since you can bundle functionality along
  with data in a class

  2. All editors provide auto-complete for class attributes, no need to
  remember dict-keys or tuple orders and stuff

  3. You can make class attributes "selectively-mutable" which means you
  can process inputs within the class but the same attributes cannot be
  tampered with outside of the class

> Notes

Perlis's Epigrams spotDL#9:  It is better to have 100 functions operate on one
data structure than 10 functions on 10 data structures.